+++ packages/authhero/test/authentication-flows/silent.spec.ts
@@ -9,11 +9,20 @@ describe("silent", () => {
     const oauthClient = testClient(oauthApp, env);
 
     const session = await env.data.sessions.create("tenantId", {
-      session_id: "sessionId",
+      id: "sessionId",
       user_id: "email|userId",
       client_id: "clientId",
-      expires_at: new Date(Date.now() + 1000).toISOString(),
       used_at: new Date().toISOString(),
+      device: {
+        last_ip: "",
+        initial_ip: "",
+        last_user_agent: "",
+        initial_user_agent: "",
+        initial_asn: "",
+        last_asn: "",
+      },
+      expires_at: new Date(Date.now() + 3600 * 1000).toISOString(),
+      clients: [],
     });
 
     const response = await oauthClient.authorize.$get(
@@ -29,7 +38,7 @@ describe("silent", () => {
       {
         headers: {
           origin: "https://example.com",
-          cookie: `tenantId-auth-token=${session.session_id}`,
+          cookie: `tenantId-auth-token=${session.id}`,
         },
       },
     );
@@ -37,5 +46,13 @@ describe("silent", () => {
     expect(response.status).toEqual(200);
     const htmlBody = await response.text();
     expect(htmlBody).toContain("access_token");
+
+    // Check that the session was updated
+    const updatedSession = await env.data.sessions.get("tenantId", "sessionId");
+    if (!updatedSession) {
+      throw new Error("Session not found");
+    }
+
+    expect(updatedSession.used_at).not.toEqual(session.used_at);
   });
 });