+++ packages/kysely/src/sessions/remove.ts
@@ -4,11 +4,9 @@ import { Database } from "../db";
 export function remove(db: Kysely<Database>) {
   return async (tenant_id: string, session_id: string): Promise<boolean> => {
     const results = await db
-      .updateTable("sessions")
-      .set({ deleted_at: new Date().toISOString() })
+      .deleteFrom("sessions")
       .where("tenant_id", "=", tenant_id)
-      .where("sessions.session_id", "=", session_id)
-      .where("sessions.deleted_at", "is", null)
+      .where("sessions.id", "=", session_id)
       .execute();
 
     return !!results.length;